requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.4
django-colorfield==0.4.1
django-crispy-forms==1.11.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
docopt==0.6.2
drf-yasg==1.20.0
gunicorn==20.1.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==20.9
Pillow==8.2.0
psycopg2==2.8.6
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.2.0
yarg==0.1.9
django-filter==2.4.0
django-tinymce==3.3.0